requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
bcrypt
passlib
pyjwt
